finishing Sipegese

+++ resources/views/admin/halaman/CrudAkunAdmin.blade.php
@@ -1,27 +1,24 @@
-@extends('pemilik.layout.TampilanPemilik')
+@extends('admin.layout.TampilanAdmin')
 @section('content')
 <!-- Container Fluid-->
         <div class="container-fluid" id="container-wrapper">
           <div class="d-sm-flex align-items-center justify-content-between mb-4">
             <h1 class="h3 mb-0 text-gray-800">Data Akun Admin</h1>
             <ol class="breadcrumb">  
-              <li class="breadcrumb-item"><a href="{{url ('dashboardadmin')}}">Dashboard</a></li>
+              <li class="breadcrumb-item"><a href="{{url ('DashboardAdmin')}}">Dashboard</a></li>
               <li class="breadcrumb-item active" aria-current="page">Data Akun Admin</li>
             </ol>
           </div>
           <hr>
-          @if($errors->any())
-                <div class="alert alert-danger alert-dismissible" role="alert">
-                            <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-                              <span aria-hidden="true">&times;</span>
-                            </button>
-                            <ul>
-                              @foreach ($errors->all() as $error)
-                              <li>{{$error}}</li>
-                              @endforeach
-                            </ul>
-                      </div>
-                    @endif
+        @if(\Session::has('alert-success'))
+                    <div class="alert alert-success alert-dismissible" role="alert">
+                        <button type="button" class="close" data-dismiss="alert" aria-label="Close">
+                          <span aria-hidden="true">&times;</span>
+                        </button>
+                        <h6><i class="fas fa-sign-out-alt"></i><b> Success!!</b></h6>
+                        {{Session::get('alert-success')}}
+                    </div>
+                  @endif
 
           <!-- DataTable with Hover -->
             <div class="col-lg-12">
@@ -29,16 +26,14 @@
                 <div class="card-header py-3 d-flex flex-row align-items-center justify-content-between">
                   
                   <a href="{{url ('TambahAkunAdmin')}}" class="btn btn-success" >Tambah data</a>
-                  <h6 class="m-0 font-weight-bold text-primary">DataTables with Hover</h6>
                 </div>
                 <div class="table-responsive p-3">
                   <table class="table align-items-center table-flush table-hover" id="dataTableHover">
                     <thead class="thead-light">
                       <tr>
-                        <th>Id</th>
+                        <th>No</th>
                         <th>Nama</th>
                         <th>Username</th>
-                        <th>Password</th>
                         <th>Email</th>
                         <th>No Handphone</th>
                         <th>Foto</th>
@@ -46,35 +41,25 @@
                         <th></th>
                       </tr>
                     </thead>
-                    <tfoot>
-                      <tr>
-                        <th>Id</th>
-                        <th>Nama</th>
-                        <th>Username</th>
-                        <th>Password</th>
-                        <th>Email</th>
-                        <th>No Handphone</th>
-                        <th>Foto</th>
-                        <th>Level</th>
-                        <th></th>
-                      </tr>
-                    </tfoot>
                     <tbody>
+                       @php
+                        $no=1;
+                      @endphp
                       @foreach($datas as $tampil)
                       <tr>
-                        <td>{{$tampil->id_admin}}</td>
+                        <td>{{$no++}}</td>
                         <td>{{$tampil->nama_admin}}</td>
                         <td>{{$tampil->username}}</td>
-                        <td>{{$tampil->password}}</td>
                         <td>{{$tampil->email}}</td>
                         <td>{{$tampil->no_hp}}</td>
-                        <td>{{$tampil->foto}}</td>
+                        <td>
+                          <img width="150px" src="{{ url('img/profil_admin/'.$tampil->foto) }}"></td>
                         <td>{{$tampil->level}}</td>
                         <td>
                           <a href="EditAkunAdmin{{$tampil->id_admin}}" class="btn btn-warning">
                             <i class="fas fa-pencil-alt"></i>
                           </a>
-                          <a href="HapusAkunAdmin{{$tampil->id_admin}}" class="btn btn-danger">
+                          <a href="HapusAkunAdmin{{$tampil->id_admin}}" class="btn btn-danger" onclick="return confirm('Anda yakin mau menghapus item ini ?')">
                             <i class="fas fa-trash"></i>
                           </a>
                         </td>
